backend/requirement.txt
fastapi 
uvicorn 
google-cloud-aiplatform 
google-cloud-storage 
google-cloud-firestore 
langchain 
langchain-google-vertexai 
pydantic 
pydantic-settings 
python-dotenv 
beautifulsoup4 
requests 
aiofiles 
python-multipart 
aiohttp 
httpx # httpx often used by langchain async
